add EdDSA support to BouncyCastleProvider
This commit adds EDDSA to the algorithms provided by the
BouncyCastleProvider.

To keep codebases compatible, this commit tries to leave the classes
imported into the `org.bouncycastle.jcajce.provider.asymmetric` package
as untouched as possible. The only change is addition of the
`DigestSignatureSpi` class, which implements a SignatureSpi for EdDSA.