treat ecc keys with unknown curves as opaque keys
Using OIDs unknown to bouncycastle resulted in an NPE in
PGPPublicKey:89, when the curve lookup returns null and the fieldsize is
accessed.

This is not necessarily the nicest solution, we could also add a more
"unknown curve" state to the regular ECC public key packets - however,
that would then lead to a situation where these keys can sometimes be
used, and sometimes not. This is the simpler fix for now, with only
*slightly* worse error reporting, since it gives the same "unknown key
algorithm" error as for an unknown key algorithm number.