revert PGPPublicKey.isMasterKey logic to distinguish by presence of subkey binding signature
This partially reverts 4bc3fca55c14f7985076ca5846a4a3343a1691f2, which
changed the logic from "true if this is a master key" to "true if this
could be a master key".

I don't know what the purpose of this change was, but it does not align
with the assumptions we make about this method in OpenKeychain at all,
consequently breaking all sorts of things downstream.